Created repositories and services. Controllers under creation

+++ src/main/java/cs/home/health/domain/MaritalStatus.java
@@ -27,7 +27,7 @@ public class MaritalStatus implements Serializable {
 	@Id
 	private Long id;
 
-	private Long code;
+	private Integer code;
 
 	private String title;
